refactor(#223): narrow IQfcFormViewer to intent-level seams for testability
Implements the approved atomic plan for issue #223, making the QuickFiler
WinForms form testable via Passive-View seams while keeping Form-derived
and Designer code [ExcludeFromCodeCoverage].

- Seam A: extract QfcFormKeyHandler.IsAltKeyCommand (pure, 100% covered);
  call it from QfcFormViewer/Dark/Expanded ProcessCmdKey; add
  [ExcludeFromCodeCoverage] to the Dark/Expanded variants.
- Seam B: replace 5 raw control properties on IQfcFormViewer with command
  events (Ok/Cancel/Undo/Skip/ItemsPerLoadValueChanged) and state
  properties (SkipButtonText/Enabled, ItemsPerLoadValue/Enabled).
- Seam C: add SwapItemTableLayout; narrow L1v0L2L3v_TableLayout to get-only;
  ActivateQueuedTlp swaps via the new method.
- Seam D: add CaptureTlpCellStates/GetKeyEventExclusionControls/
  ItemViewerTemplateMargin; remove the two template properties; refactor
  CaptureItemSettings and the keyboard-exclusion call sites.
- Phase 0: split QfcFormController.cs (1142 lines) into 4 partial classes,
  all <500 lines, to satisfy the file-size cap before adding code.

Consumers updated in-repo (QfcFormController, QfcCollectionController,
QfcHomeController). New seam tests verify command-event routing, skip-flow
state, and CaptureItemSettings paths via Moq Raise/VerifySet. Toolchain
passes (csharpier, analyzers, nullable/TWAE, 196/196 MSTest with coverage).

Pre-existing-debt dispositions: QfcCollectionController.cs (2296 lines, not
split, net-negative edit) and QfcFormControllerTests.cs (821 lines,
net-neutral; new tests routed to QfcFormControllerSeamTests.cs).

Refs #223
