+++ docs/features/active/2026-06-28-qfc-form-viewer-testability-223/evidence/baseline/baseline-analyzers.2026-06-28T20-52.md
@@ -1,0 +1,7 @@
+# Baseline — Analyzer Build (Issue #223)
+
+Timestamp: 2026-06-28T20-52
+Command: msbuild TaskMaster.sln -t:Build -p:Configuration=Debug -p:Platform="Any CPU" -p:EnableNETAnalyzers=true -p:EnforceCodeStyleInBuild=true
+EXIT_CODE: 0
+
+Output Summary: Build succeeded. 0 Error(s), 68 Warning(s). All warnings are pre-existing and outside the issue-#223 scope: CS8632 (nullable annotation outside #nullable context) and CS0067 (unused event) in test projects (TaskMaster.Test, UtilitiesCS.Test). No analyzer errors. NuGet restore (168 packages) was required first on this fresh worktree. Diagnostic headline: 0 errors / 68 warnings baseline; later phases must not introduce new analyzer errors.
