+++ docs/features/active/2026-06-28-qfc-form-viewer-testability-223/evidence/baseline/baseline-csharpier.2026-06-28T20-52.md
@@ -1,0 +1,7 @@
+# Baseline — CSharpier Format Check (Issue #223)
+
+Timestamp: 2026-06-28T20-52
+Command: dotnet tool run csharpier check .
+EXIT_CODE: 0
+
+Output Summary: Checked 1183 files in ~3.5s. No format drift detected on the clean tree (exit 0). Baseline format state is clean; any later format failure is attributable to this cycle's edits.
