+++ docs/features/active/2026-06-28-qfc-form-viewer-testability-223/evidence/baseline/phase0-instructions-read.md
@@ -1,0 +1,26 @@
+# Phase 0 — Instructions Read (Issue #223)
+
+Timestamp: 2026-06-28T20-52
+
+Policy Order:
+1. CLAUDE.md (standing project instructions; always loaded)
+2. .claude/rules/general-code-change.md (cross-language code change policy)
+3. .claude/rules/general-unit-test.md (cross-language unit test policy)
+4. .claude/rules/csharp.md (C#-specific toolchain and standards)
+
+Files read (policy):
+- C:\Users\DanMoisan\repos\TaskMaster-wt-2026-06-28-18-50\CLAUDE.md
+- C:\Users\DanMoisan\repos\TaskMaster-wt-2026-06-28-18-50\.claude\rules\general-code-change.md
+- C:\Users\DanMoisan\repos\TaskMaster-wt-2026-06-28-18-50\.claude\rules\general-unit-test.md
+- C:\Users\DanMoisan\repos\TaskMaster-wt-2026-06-28-18-50\.claude\rules\csharp.md
+- C:\Users\DanMoisan\repos\TaskMaster-wt-2026-06-28-18-50\.claude\rules\ci-workflows.md
+- C:\Users\DanMoisan\repos\TaskMaster-wt-2026-06-28-18-50\.claude\rules\tonality.md
+
+Files read (authoritative inputs):
+- docs/features/active/2026-06-28-qfc-form-viewer-testability-223/spec.md
+- docs/features/active/2026-06-28-qfc-form-viewer-testability-223/issue.md
+- artifacts/research/2026-06-28T18-00-qfc-form-viewer-testability-research.md
+- artifacts/research/2026-06-28T19-00-qfc-seam-c-d-implementation-research.md
+- docs/features/active/2026-06-28-qfc-form-viewer-testability-223/plan.2026-06-28T20-20.md
+
+Output Summary: All four policy files plus CI/tonality rules read in the required order; all four authoritative inputs and the plan-of-record read. Work Mode confirmed full-feature (spec.md + issue.md authoritative). Toolchain order confirmed: csharpier -> analyzers msbuild -> nullable/TreatWarningsAsErrors msbuild -> vstest with coverage. Evidence path invariant confirmed: docs/features/active/2026-06-28-qfc-form-viewer-testability-223/evidence/<kind>/.
