+++ docs/features/active/2026-06-28-qfc-form-viewer-testability-223/evidence/qa-gates/final-csharpier.2026-06-28T20-52.md
@@ -1,0 +1,7 @@
+# Final QC — CSharpier (Issue #223)
+
+Timestamp: 2026-06-28T20-52
+Command: dotnet tool run csharpier check .
+EXIT_CODE: 0
+
+Output Summary: Checked 1189 files, exit 0. No format drift on the final tree. No files changed by the check, so the loop did not need to restart.
