+++ docs/features/active/2026-06-28-qfc-form-viewer-testability-223/evidence/qa-gates/final-nullable.2026-06-28T20-52.md
@@ -1,0 +1,7 @@
+# Final QC — Nullable / TreatWarningsAsErrors Build (Issue #223)
+
+Timestamp: 2026-06-28T20-52
+Command: msbuild TaskMaster.sln -t:Build -p:Configuration=Debug -p:Platform="Any CPU" -p:Nullable=enable -p:TreatWarningsAsErrors=true
+EXIT_CODE: 0
+
+Output Summary: Build succeeded. 0 Warning(s), 0 Error(s). Nullable gate clean (AC7).
