+++ docs/features/active/2026-06-28-qfc-form-viewer-testability-223/evidence/qa-gates/final-tests-coverage.2026-06-28T20-52.md
@@ -1,0 +1,11 @@
+# Final QC — Tests + Coverage (Issue #223)
+
+Timestamp: 2026-06-28T20-52
+Command: vstest.console.exe QuickFiler.Test\bin\Debug\QuickFiler.Test.dll /EnableCodeCoverage /InIsolation
+EXIT_CODE: 0
+
+Output Summary:
+- Total tests: 196. Passed: 196. Failed: 0 (AC7).
+- Post-change QfcFormController line coverage (filename+line keyed across all 4 partials): 363/700 = 51.86%.
+- QfcFormKeyHandler (new code) coverage: 2/2 = 100.0%.
+- Process-wide line coverage (QuickFiler.Test run): 12.86% (9800/76203) — consistent single-assembly reference metric (instruments all loaded modules; not the repo-wide first-party gate).
