+++ docs/features/active/2026-06-28-qfc-form-viewer-testability-223/evidence/qa-gates/p1-analyzers.2026-06-28T20-52.md
@@ -1,0 +1,7 @@
+# Phase 1 — Analyzer Build (Issue #223)
+
+Timestamp: 2026-06-28T20-52
+Command: msbuild TaskMaster.sln -t:Build -p:Configuration=Debug -p:Platform="Any CPU" -p:EnableNETAnalyzers=true -p:EnforceCodeStyleInBuild=true
+EXIT_CODE: 0
+
+Output Summary: Build succeeded. 0 Error(s), 47 Warning(s). No new analyzer errors versus baseline (baseline was 0 errors). Warning count differs from baseline 68 only due to incremental-recompile scope (pre-existing CS8632/CS0067 in test projects not re-emitted); no new diagnostics introduced by the partial-class split.
