+++ docs/features/active/2026-06-28-qfc-form-viewer-testability-223/evidence/qa-gates/p1-csharpier.2026-06-28T20-52.md
@@ -1,0 +1,7 @@
+# Phase 1 — CSharpier (Issue #223)
+
+Timestamp: 2026-06-28T20-52
+Command: dotnet tool run csharpier format <4 split files> ; then dotnet tool run csharpier check .
+EXIT_CODE: 0
+
+Output Summary: Formatted 4 split files; repo-wide check passed (Checked 1186 files, exit 0). No unresolved format drift. The 3 new partial files compile-formatted cleanly.
