+++ docs/features/active/2026-06-28-qfc-form-viewer-testability-223/evidence/qa-gates/p1-file-sizes.2026-06-28T20-52.md
@@ -1,0 +1,12 @@
+# Phase 1 — Partial-Class File Sizes (Issue #223)
+
+Timestamp: 2026-06-28T20-52
+Command: wc -l on QfcFormController.cs and its three new partials (post-csharpier)
+
+Line counts:
+- QuickFiler/Controllers/QfcFormController.cs: 195 lines (was 1142; retains usings, namespace/class decl, Constructors, Private Variables, Public Properties)
+- QuickFiler/Controllers/QfcFormController.SetupDisposal.cs: 298 lines (Setup and Disposal region)
+- QuickFiler/Controllers/QfcFormController.EventHandlers.cs: 399 lines (Event Handlers region)
+- QuickFiler/Controllers/QfcFormController.Actions.cs: 311 lines (Major Actions region)
+
+Output Summary: All four files are < 500 lines (195 / 298 / 399 / 311). Counts align with plan expectations (~190 / ~286 / ~387 / ~299). Pure structural split; no method bodies changed. AC6 satisfied for the QfcFormController split.
