+++ docs/features/active/2026-06-28-qfc-form-viewer-testability-223/evidence/qa-gates/p1-tests-coverage.2026-06-28T20-52.md
@@ -1,0 +1,9 @@
+# Phase 1 — Tests + Coverage (Issue #223)
+
+Timestamp: 2026-06-28T20-52
+Command: vstest.console.exe QuickFiler.Test\bin\Debug\QuickFiler.Test.dll /EnableCodeCoverage /InIsolation
+EXIT_CODE: 0
+
+Output Summary:
+- Total tests: 181. Passed: 181. Failed: 0. Passing count EQUALS the P0-T5 baseline (181) — confirms the partial-class split caused no test change.
+- Process-wide line coverage (QuickFiler.Test run): 12.52% (lines-covered 9524 / lines-valid 76066) — IDENTICAL to the P0-T5 baseline, confirming a pure structural refactor with zero coverage movement.
