+++ docs/features/active/2026-06-28-qfc-form-viewer-testability-223/evidence/qa-gates/p2-csharpier.2026-06-28T20-52.md
@@ -1,0 +1,7 @@
+# Phase 2 — CSharpier (Issue #223)
+
+Timestamp: 2026-06-28T20-52
+Command: dotnet tool run csharpier format <5 Seam A files> ; then dotnet tool run csharpier check .
+EXIT_CODE: 0
+
+Output Summary: Formatted 5 files; repo-wide check passed (Checked 1188 files, exit 0). No unresolved format drift.
