+++ docs/features/active/2026-06-28-qfc-form-viewer-testability-223/evidence/qa-gates/p2-tests-coverage.2026-06-28T20-52.md
@@ -1,0 +1,10 @@
+# Phase 2 — Tests + Coverage (Seam A) (Issue #223)
+
+Timestamp: 2026-06-28T20-52
+Command: vstest.console.exe QuickFiler.Test\bin\Debug\QuickFiler.Test.dll /EnableCodeCoverage /InIsolation
+EXIT_CODE: 0
+
+Output Summary:
+- Total tests: 185. Passed: 185. Failed: 0. (Baseline 181 + 4 new QfcFormKeyHandler tests; all prior tests still pass.)
+- New tests all PASS: IsAltKeyCommand_WithAltKey_ReturnsTrue, IsAltKeyCommand_WithAltPlusOtherKey_ReturnsTrue, IsAltKeyCommand_WithControlKey_ReturnsFalse, IsAltKeyCommand_WithNone_ReturnsFalse.
+- QfcFormKeyHandler line coverage: 100.0% (2/2 lines covered) — exceeds the AC5 new-code >= 90% floor.
