+++ docs/features/active/2026-06-28-qfc-form-viewer-testability-223/evidence/qa-gates/p3-analyzers.2026-06-28T20-52.md
@@ -1,0 +1,7 @@
+# Phase 3 — Analyzer Build (Issue #223)
+
+Timestamp: 2026-06-28T20-52
+Command: msbuild TaskMaster.sln -t:Build -p:Configuration=Debug -p:Platform="Any CPU" -p:EnableNETAnalyzers=true -p:EnforceCodeStyleInBuild=true
+EXIT_CODE: 0
+
+Output Summary: Build succeeded. 0 Error(s), 47 Warning(s) (all pre-existing CS8632/CS0067 in test projects). A targeted grep for QuickFiler-scope warnings/errors excluding the CS8632/CS0067 baseline returned empty — no new analyzer diagnostics from Seams B/C/D.
