+++ docs/features/active/2026-06-28-qfc-form-viewer-testability-223/evidence/qa-gates/p3-file-sizes.2026-06-28T20-52.md
@@ -1,0 +1,25 @@
+# Phase 3 — File Sizes (Issue #223)
+
+Timestamp: 2026-06-28T20-52
+Command: wc -l on all production and test files modified/created in Phases 1-3
+
+Production files:
+- QuickFiler/Interfaces/IQfcFormViewer.cs: 51 (< 500)
+- QuickFiler/Viewers/QfcFormViewer.cs: 262 (< 500)
+- QuickFiler/Viewers/QfcFormViewerDark.cs: 55 (< 500)
+- QuickFiler/Viewers/QfcFormViewerExpanded.cs: 55 (< 500)
+- QuickFiler/Controllers/QfcFormController.cs: 195 (< 500)
+- QuickFiler/Controllers/QfcFormController.SetupDisposal.cs: 232 (< 500)
+- QuickFiler/Controllers/QfcFormController.EventHandlers.cs: 399 (< 500)
+- QuickFiler/Controllers/QfcFormController.Actions.cs: 311 (< 500)
+- QuickFiler/Controllers/QfcFormKeyHandler.cs: 20 (< 500)
+- QuickFiler/Controllers/QfcHomeController.cs: 454 (< 500)
+- QuickFiler/Controllers/QfcCollectionController.cs: 2296 (DISPOSITIONED pre-existing production debt; P0-T6 baseline 2299; net -3 from Seam C `ActivateQueuedTlp` rewrite; NOT split; <= baseline)
+
+Test files:
+- QuickFiler.Test/Controllers/QfcFormControllerTests.cs: 821 (DISPOSITIONED pre-existing test-cap debt; P0-T6 baseline 823; in-place Seam B migration is net-neutral, NOT increased — actually -2; no new [TestMethod] cases added here)
+- QuickFiler.Test/Controllers/QfcFormControllerSeamTests.cs: 326 (NEW; < 500; holds all 11 new seam tests)
+- QuickFiler.Test/Controllers/QfcHomeControllerRunAsyncTests.cs: 446 (< 500)
+- QuickFiler.Test/Controllers/QfcFormKeyHandlerTests.cs: 67 (< 500)
+
+Output Summary: Every modified production file except the dispositioned QfcCollectionController.cs is < 500 lines. QfcCollectionController.cs is net-negative (2296 <= 2299 baseline), recorded as pre-existing-debt disposition (AC6). QfcFormControllerTests.cs is not increased versus its 823 baseline (821 <= 823). New QfcFormControllerSeamTests.cs is < 500 (326). AC6 satisfied.
