+++ docs/features/active/2026-06-28-qfc-form-viewer-testability-223/evidence/qa-gates/p3-tests-coverage.2026-06-28T20-52.md
@@ -1,0 +1,12 @@
+# Phase 3 — Tests + Coverage (Seams B/C/D) (Issue #223)
+
+Timestamp: 2026-06-28T20-52
+Command: vstest.console.exe QuickFiler.Test\bin\Debug\QuickFiler.Test.dll /EnableCodeCoverage /InIsolation
+EXIT_CODE: 0
+
+Output Summary:
+- Total tests: 196. Passed: 196. Failed: 0. (Baseline 181 + 4 KeyHandler (P2) + 11 new seam tests (P3). All prior tests still pass after the Seam B migrations.)
+- 11 new seam tests pass: RegisterFormEventHandlers_WiresAllIntentCommandEvents, RegisterFormEventHandlers_UsesExclusionControlsFromFormViewer, OkClicked/CancelClicked/UndoClicked/ItemsPerLoadValueChanged/SkipClicked routing, ButtonSkipHandler skip-flow, CaptureItemSettings populated/null/null-RowStyles.
+- QfcFormController coverage (keyed by filename+line across all 4 partials): 363/700 = 51.86%, up from baseline 301/767 = 39.24%. No coverage regression on changed lines; the denominator dropped because Seam D moved the ~58-line TlpCellStates construction block into the [ExcludeFromCodeCoverage] Form.
+- QfcFormKeyHandler: 2/2 = 100%.
+- Process-wide line coverage (QuickFiler.Test run): 12.86% (9800/76203) — consistent single-assembly reference metric.
